Charts: Enable Hourly and half-hourly data
We use the Simtricity flow API (Google visualisation version, as used by the Simtricity charts)
to fetch hourly and half-hourly generation data.  As this data comes back in kW rather than
kWh, the hourly data remains the same when converting, but the half-hourly data needs to
be halved to get the kWh value rather than the kW value.

The flow API requires a flow token.  We take this from the back-end database at
bec-monitoring as we don't know how to get it directly from the Simtricity APIs!

Although the data from the flow API comes back as a JSON object, it's not in the structure
needed by our ChartService.  We add a GVIZToObjects function to the UtilitiesService
which parses it as a string to create the array of objects in the form required.

TODO: When resolution is changed, alter tooltips and ticks to show time rounded
to the nearest hour or half hour (depending on resolution).