Issue #15: Headline: Correct number of days if less than requested
The DataService.js file stores the duration in days as part of it's metadata.  It was just
using the difference between the requested start and end dates, but we can instead
store it after reading data by using the difference between the time in the first and last
pieces of data returned.